Update README.md
MQTT PROTOCOL AND IOT CLOUD SERVICE
•	Reason
The Internet of Things is emerging as the third wave in the development of the internet. Internet of things (IoT) is expected to have a massive impact on consumer products, business and widerculture, but these are still early days. Given its potential for very wide applicability to almost all verticals and aspects of business, industries, manufacturing, consumer goods, supply chains, etc..IoT as a whole is very broad area. This report focuses specifically to its adoption to Smart Home.
•	Problem & Methods
With the goal of applying IoT to life. We decided to implement the project of "RESEARCHING MQTT PROTOCOL AND IOT CLOUD SERVICE". The project focuses on establishing device communication via MQTT protocol and extending the monitoring and control of devices via website interface.
•	Results & Conclusion
In addition, the project also applies IoT solutions for building websites that independently control sensor devices. Through the web we can control on / off lights, fans, read the temperature-humidity sensor values, light warning when gas is detected.